Update README.md
## Batch Macro used in the "Superstore" Project.
This [batch macro](Shortest%20Route%20Batch.yxmc) uses batch process to find the nearest location using spatial analytics. This macro aims to solve the "Traveling Salesman Problem", and it is used to produce many outputs using batches. It needs an upstream macro, that will actually find the most optimal route, using MIN function, which is tackled further in the "Shortest Route Checker" Standard Macro.

### Alteryx Custom Made Macro Group
- Using Custom Made Batch Macro
**![CustomMacro2.png](Workflow-Screens/CustomMacro2.png)**

## Standard Macro used in "Superstore" project.
This [standard macro](Shortest%20Route%20Checker.yxmc) feeds iterative and batch process to find the nearest location using spatial analytics. This macro aims to solve the "Traveling Salesman Problem", but it's limitation is the longer run times, depending on the amount of locations it needs to connect and check.

### Alteryx Custom Made Macro Group
- Using Custom Made Standard Macro
**![CustomMacro.png](Workflow-Screens/CustomMacro.png)**